Add MacOS and Windows stubs.
